pyproject.toml
[project]
name = "piano-teacher"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "mido>=1.3.3",
    "pyfluidsynth==1.3.4",
    "pygame>=2.6.1",
    "python-rtmidi==1.5.8",
]
